Revive `EntityUtil.merge()` function.
